backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
boto3
python-dotenv
Pillow
joblib
scikit-learn
pandas
psutil
# Greengrass v2 IPC SDK -- only required when USE_GREENGRASS_IPC=true
awsiotsdk>=1.21.0
